+++ mfio-rt/src/native/impls/iocp/mod.rs
@@ -274,7 +274,7 @@ impl OperationMode {
         match self {
             Self::FileRead(pkt, buf) => match res {
                 Ok(read) if (read as u64) < pkt.len() => {
-                    let (left, right) = pkt.split_at(read as u64);
+                    let (left, right) = pkt.split_at(read);
                     if let Err(pkt) = left {
                         assert!(!buf.0.is_null());
                         let buf = unsafe { &*buf.0 };
@@ -299,7 +299,7 @@ impl OperationMode {
             },
             Self::FileWrite(pkt, _) => match res {
                 Ok(read) if (read as u64) < pkt.len() => {
-                    let (left, right) = pkt.split_at(read as u64);
+                    let (left, right) = pkt.split_at(read);
                     deferred_pkts.ok(left);
                     right.error(io_err(State::Nop));
                 }
